requirements.txt
raygun4py==4.3.0
sqlalchemy==1.3.22
discord.py==1.6.0